Remove isStaticFile function to streamline middleware code and eliminate redundancy
